+++ src/backend/core/tasks/__init__.py
@@ -25,6 +25,7 @@ from core.tasks import file as _file
 from core.tasks import meeting_sessions as _meeting_sessions
 from core.tasks import member_import as _member_import
 from core.tasks import offboarding as _offboarding
+from core.tasks import rooms as _rooms
 from core.tasks import summary as _summary
 from core.tasks import summary_exports as _summary_exports
 from core.tasks import summary_notifications as _summary_notifications